Edited in accordance with second pull-request
